requirements.txt
Flask
tensorflow
numpy
pillow
gunicorn
